minor css improvements and bugfixes
Adapt styled components to new version of sanitize.css
Align userItems children vertical
Move StyledFormControl to a separate file for code reusability
