+++ app/components/FortuneWheelRaffle/DialogRoot.js
@@ -9,6 +9,7 @@ const DialogRoot = styled.div`
   bottom: 0;
   left: 0;
   right: 0;
+  z-index: 1;
 `;
 
 export default DialogRoot;
