+++ app/components/Panel/PanelTitle.js
@@ -2,6 +2,7 @@ import styled from 'styled-components';
 
 const PanelTitle = styled.h2`
   color: ${(props) => props.theme.staticTextColor};
+  line-height: 1;
   margin: 5px 10px;
 `;
