+++ app/components/StyledTextField/StyledFormControl.js
@@ -1,0 +1,29 @@
+import styled from 'styled-components';
+import FormControl from '@material-ui/core/FormControl';
+
+const StyledFormControl = styled(FormControl)`
+  width: 100%;
+  input {
+    color: ${(props) => props.theme.staticTextColor};
+  }
+  svg {
+    color: ${(props) => props.theme.staticTextColor};
+  }
+  label span {
+    color: ${(props) => props.theme.inputLabel};
+  }
+  label.Mui-focused span {
+    color: ${(props) => props.theme.color};
+  }
+  .MuiInput-input {
+    color: ${(props) => props.theme.staticTextColor};
+  }
+  .MuiInput-underline:hover:not(.Mui-disabled):before {
+    border-bottom: 2px solid ${(props) => props.theme.color};
+  }
+  .MuiInput-underline:after {
+    border-bottom: 2px solid ${(props) => props.theme.color};
+  }
+`;
+
+export default StyledFormControl;
