+++ app/components/WelcomeDialog/CookieConsent.js
@@ -33,6 +33,7 @@ const Image = styled.img`
 `;
 
 const Description = styled.span`
+  line-height: 1.1;
   max-width: 30vw;
   @media (orientation: portrait) {
     max-width: 58vw;
